Updated way of deleting a match
Process is that we delete all elo histories
that are related to the match, and then the
matches that happen after that match.  Then
we can restore the  ELO rankings for  every
player who is affected to their  former ELO
score. Now it is safe to delete the  result
from the match. After that is done, we take
every result and calculate ELO rankings for
each player again, using the new scores  as
a base for calculations.