PlayerStatisticsCollection added
PlayerStatisticsCollection keeps track of all the PlayerStatistics. It
can be sorted by various parameters and then looped over in a foreach.
The statistics service is then responsible for populating the collection
with some verbose queries, but this developer is not quite sure how to
make it more efficient or more pretty. Make it work first though :)

Upside is though, that we only have 4 queries to get all statistics for
all players. It can be debated if there should be 4 queries per
statistic (such as goal scored and games won) that could possibly be
optimized to actually be fewer instead of doin sorting in PHP.