Filtering out players  with too few games
When players have played fewer games than the minimum games included
constant, they are not considered in the sorted collection. Maybe they
should be removed from the player collection, but that would maybe be
premature if user decides to add more players later. This could be moved
over to paramteres.yml or a modifieable constant saved in the database
by the administrators.

Having it in the sort instead of add, is better because it is the total
minimum amount games played, not the minimum amount of games per
addition to the collection.