Switching out statistic services and materializing
Materialized the statistics page and changed the statistics provider to
the new one. This cuts the time spent on DB queries to 50ms from 235ms
on average. And from about 280 DB queries to 35 queries. Most
importantly that remains a constant when before it was linear in terms
of players (and perhaps matches). Those results were based off 18
players and about 500 results.